切换统一绩效模型并收缩旧结构（Ticket 21）
完成 contract 门禁与旧表删除，统一模板、提交、阶段结果和结果版本读写；补齐 D/M、整体退回、权限隐私及分类监控的真实 PostgreSQL 旅程。

+++ backend/test/active-cycle-rollback-postgres.integration-spec.ts
@@ -122,7 +122,7 @@ describe('Ticket 17 PostgreSQL 周期整体退回集成', () => {
     await first.query(
       `
       INSERT INTO "${schema}"."notification_events" (dedupe_key, rollback_id, receiver_open_id)
-      SELECT 'result-invalidated:' || $1 || ':' || participant.id, $1, participant.id::text
+      SELECT 'result-invalidated:' || $1::integer || ':' || participant.id, $1::integer, participant.id::text
       FROM "${schema}"."participants" AS participant
       WHERE participant.id IN (102, 103, 104, 105)
     `,
